+++ SmartRacecar/src/main/java/be/uantwerpen/fti/ds/sc/smartracecar/manager/Location.java
@@ -2,10 +2,10 @@ package be.uantwerpen.fti.ds.sc.smartracecar.manager;
 
 class Location {
 
-    private Long vehicleID;
-    private int wayPointID;
+    private long vehicleID;
+    private long wayPointID;
 
-    Location(Long vehicleID, int wayPointID){
+    Location(long vehicleID, long wayPointID){
         this.vehicleID = vehicleID;
         this.wayPointID = wayPointID;
     }
